Comment-out exception reporting for now
This seems to cause strange errors
