+++ src/device/quirks.jl
@@ -2,7 +2,7 @@
 
 macro print_and_throw(description)
     quote
-        @errprintf($description)
+        # @errprintf($description)
         throw(nothing)
     end
 end
